Importing Routes from the Kernel
Back in the test, thanks to FrameworkBundle, we can use a really cool trait to make life simpler. Full disclosure, I helped created the trait - so of course I think it's cool. But really, it makes life easier: use MicroKernelTrait. Remove registerContainerConfiguration() and, instead go back again to the Code -> Generate menu - or Command + N on a Mac - and implement the two missing methods: configureContainer(), and configureRoutes().
Cool! So... let's import our route! $ routes->import(), then the path to that file: __DIR__.'/../../src/Resources/config/routes.xml'.

Setting up the Test Client

Nice! And... that's really all the kernel needs. Back up in testIndex(), create the new kernel: new KnpULoremIpsumControllerKernel().

Now, you can almost pretend like this a normal functional test in a normal Symfony app. Create a test client: $ client = new Client() - the one from FrameworkBundle - and pass it the .

Use this to make requests into the app with $ client->request(). You will not get auto-completion for this method - we'll find out why soon. Make a GET request, and for the URL... actually, down in configureRoutes(), ah, I forgot to add a prefix! Add /api as the second argument. Make the request to /api/.

Cool! Let's dump the response to see what it looks like: var_dump($ client->getResponse()->getContent()). Then add an assert that 200 matches $ client->getResponse()->getStatusCode().

Alright! Let's give this a try! Find your terminal, and run those tests!

./vendor/bin/simple-phpunit

Woh! They are not happy:

    Fatal error class BrowserKit\Client does not exist.

Hmm. This comes from the http-kernel\Client class. Here's what's happening: we use the Client class from FrameworkBundle, that extends Client from http-kernel, and that tries to use a class from a component called browser-kit, which is an optional dependency of http-kernel. Geez.

Basically, we're trying to use a class from a library that we don't have installed.