You know the drill, find your terminal and run: 'composer require symfony/browser-kit --dev'
When that finishes, try the test again!

./vendor/bin/simple-phpunit

Oof. It just looks awful:

    LogicException: Container extension 'framework' is not registered.

This comes from ContainerBuilder, which is called from somewhere inside MicroKernelTrait. This is a bit tougher to track down. When we use MicroKernelTrait, behind the scenes, it adds some framework configuration to the container in order to configure the router. But... our kernel does not enable FrameworkBundle!