What if a user wants to change the behavior of our controller? Symfony does have a way to override controllers from a bundle... but not if that controller is registered as a service, like our controller. Well, ok, thanks to Symfony's incredible container, there is always a way to override a service. But let's not make our users do crazy things! If someone wants to tweak how our controller behaves, let's make it easy!
How? By dispatching a custom event. Ready for our new goal? I want to allow a user to change the data that we return from our API endpoint. Specifically, we're going to add a third key to the JSON array from our app.
Custom Event Class

The first step to dispatching an event is to create an event class. Create a new Event directory with a PHP class inside: call it FilterApiResponseEvent. I just made that up.

Make this extend a core Event class from Symfony. When you dispatch an event, you have the opportunity to pass an Event object to any listeners. To be as awesome as possible, you'll want to make sure that object contains as much useful information as you can.

In this case, a listener might want to access the data that we're about to turn into JSON. Cool! Add public function __construct() with an array $ data argument. I'll press Alt+Enter and choose 'Initialize Fields' to create a data property and set it.

Then, we need a way for the listeners to access this. And, we also want any listeners to be able to set this. Go back to the Code -> Generate menu, or Command + N on a Mac, choose 'Getter and Setters' and select data.

It's ready!

Dispatching the Event

Head to your controller: this is where we'll dispatch that event. First, set the data to a  variable and then create the event object: $ event = new FilterApiResponseEvent() passing it the data. I'm not going to dispatch the event quite yet, but at the end, pass $ event->getData() to the json method.