At this point, the user can control the word provider. But, there's only ever one word provider. That may be fine, but I want to make this more flexible! And, along the way, learn about one of the most important, but complex systems that is commonly used in bundles: the tag & compiler pass system.
First, let's make our mission clear: instead of allowing just one word provider, I want to allow many word providers. I also want other bundles to be able to automatically add new word providers to the system. Basically, I want a word provider plugin system.
Allowing Multiple Word Providers

To get this started, we need to refactor KnpUIpsum: change the first argument to be an array of $ wordProviders. Rename the property to $ wordProviders, and I'll add some PHPDoc above this to help with auto-completion: this will be an array of WordProviderInterface[].
Let's also add a new property called wordList: in a moment, we'll use this to store the final word list, so that we only need to calculate it once.
The big change is down below in the getWordList() method. First, if null === $ this->wordList, then we need to loop over all the word providers to create that word list.

Once we've done, that, at the bottom, return $ this->wordList.
Back in the if, create an empty  array, then loop over $ this->wordProviders as $ wordProvider. For each word provider, set $ words to an array_merge of the words so far and $ wordProvider->getWordList().
After, we need a sanity check: if the count($ words) <= 1, throw an exception: this class only works when there are at least two words. Finally, set $ this->wordList to $ words.
Perfect! This class is now just a little bit more flexible. In config/services.xml, instead of passing one word provider, add an <argument with type='collection', them move the word provider argument inside of this.
There's no fancy plugin system yet, but things should still work. Find your browser and refresh. Great! Even the article page looks fine.