The Easy Way
So... how do we hook this all up? Well, if your bundle will only need to support Symfony 3.4 or higher, there's a super easy way. Just replace the <argument type='collection'> with <argument type='tagged' tag='knpu_ipsum_word_provider' />. This tells Symfony to find all services with this tag, and pass them as a collection. And... you'd be done!

But, if you want to support earlier versions of Symfony, or you want to know how the compiler pass system works, keep watching.