And now... it's easy! The code in a compiler pass looks a lot like the code in an extension class. At the top, add $ definition = $ container->getDefinition('knpu_lorem_ipsum.knpu_ipsum').
Ultimately, we need to modify this services's first argument. Create an empty $ references array. And, in the foreach, just add stuff to it: $ references[] = new Reference() and pass in the .

Finish this with $ definition->setArgument(), pass it 0 for the first argument, and the array of reference objects.
We're done! Go back to our browser and try it! Woohoo! We're now passing an array of all of the word provider services into the KnpUIpsum class.... which... yea, is just one right now.