Our bundle is ready to be shared with the world! So let's take care of a few last details, and publish our bundle to Packagist! Choosing a License
But, before we publish this anywhere, we need do some boring, but very important legal work. Go to choosealicense.com and find the license that works best for you. Symfony is licensed MIT, and that's definitely the best practice. Whatever you choose, copy the license, find your bundle code, and at the root, create the LICENSE file.