Adding the .travis.yml File
The standard for continuous integration of open source libraries is definitely Travis CI. And if you go back to the 'Best Practices' docs for bundles('https://symfony.com/doc/current/bundles/best_practices.html'), near the top, Symfony has an example of a robust Travis configuration file! Awesome!

Copy this entire thing, go back to the bundle, and, at the root, create a new file - .travis.yml. Paste!

We'll talk about some of the specifics of this file in a minute. But first, in your terminal, add everything we've been working on, commit, and push.
Activating Travis CI

With the Travis config file in place, the next step is to activate CI for the repo. Go to travis-ci.org and make sure you're signed in with GitHub. Click the '+' to add a new repository, I'll select the 'KnpUniversity' organization and search for lorem.

Huh. Not found. Because it's a new repository, it probably doesn't see it yet. Click the 'Sync Account' button to fix that. And... search again. There it is! If it's still not there for you, keep trying 'Sync Account': sometimes, it takes several tries.

Activate the repo, then click to view it. To trigger the first build, under 'More options', click, ah, 'Trigger build'! You don't need to fill in any info on the modal.

Oh, and from now on, a new build will happen automatically whenever you push. We only need to trigger the first build manually. And... go go go!